Create simple intrest calculator.java
Write a program to enter P,T (principal and time) and calculate

simple interest. Use the Scanner class for taking inputs from the

console. Users will enter the principal amount and time(in years).

Create an if-else statement and modify the interest rate based on the

principal amount. If the amount > 10000 then the interest rate is 10%. If the

amount is between 10000 and 5000 then make the interest rate 8%. For

any amount below 5000 the interest rate should be 5%.